Create Outreach.tsx
Started creating the outreach page on the website under components, i'm not sure if it is the correct place to put it. I used the mission statement code as well as some of the old code from the old website and the text is from the old website. Subject to be edited both code and text.